[S09P12B205-484] Feat: 온라인,오프라인 구현

See merge request s09-webmobile1-sub2/S09P12B205!433

+++ Front-End/mannamanna/src/pages/Note/Modal/DeleteNoteModal.tsx
@@ -1,5 +1,5 @@
 import { useRecoilState } from "recoil";
-import { DeleteNoteAtom, sendNoteIdAtom } from "../../../Recoil/State";
+import { DeleteNoteAtom, idAtom, sendNoteIdAtom } from "../../../Recoil/State";
 import MacBookBox from "../../../components/common/macbookBox";
 import Modal from "@mui/material/Modal";
 import React from "react";
@@ -23,10 +23,11 @@ export const DeleteNoteModal = () => {
   const handleOpen = () => setOpen(true);
   const handleClose = () => setOpen(false);
   const [noteId, setNoteId] = useRecoilState(sendNoteIdAtom);
+  const [UserId,_] =useRecoilState(idAtom);
 
   const sendDelete = async (e: React.MouseEvent<HTMLButtonElement>) => {
     try {
-      const response = await api.delete(`note/${noteId}`);
+      const response = await api.delete(`note/${noteId}/${UserId}`);
       console.log(response.data);
       console.log("잘지웠군");
     } catch (error) {
@@ -36,7 +37,7 @@ export const DeleteNoteModal = () => {
   const handleRemove = async (e: React.MouseEvent<HTMLButtonElement>) => {
     await sendDelete(e);
     handleClose();
-    // window.location.reload();
+    window.location.reload();
   };
 
   return (
